feat(AMI-1943): upgrade to Capacitor v4 (#18)
Tested it for a short while on both Android and iOS. No prominent errors
found so far.

We might want to extensively test the secure storage plugin though.
Because I found this:
https://github.com/martinkasa/capacitor-secure-storage-plugin/issues/71

I also seem to remember that the `capacitor-branch-deep-links` plugin
version `4.2.0` caused us some trouble. Not sure why anymore though. So
we probably have to do some extra digging and testing there as well.