refactor: move the chat-feature to a seperate package (#84)
Chat feature is now a whole stand-alone package with Stream integration.
 This gives us more freedom to develop the chat as well as other native
features independently and makes the app more maintainable and scalable.

The `App` project has no idea of existence of the stream SDK. And only
imports `Amigos Chat Package` The goal is (by creating our own UI) to
also get rid of `StreamChatSwiftUI` package as this tight couple the
chat client to it’s UI and we’re limited with new features we’d like to
implement.

We can also create a new Xcode project to run the chat independently

---------

Co-authored-by: Vissie <19539422+Vissie2@users.noreply.github.com>