chore: add SwiftLint package to `Amigos Chat Package` (#91)
Two main changes here: 

1. 
SwiftLint is now a plugin part of the `Amigos Chat Package` so we can
lint the package it self and anything outside of the package. This was
far the easiest and most straightforward thing to do. If we move to
different native features, we can consider it adding the plugin to Xcode
projects instead of the packages.

2. 
Everything where SwiftLint was disabled is carefully refactored. I've
tried to explain the commits clearly why it's refactored and what
SwiftLint rule it violated.