refactor(AMI-2754): implement new message view structure (#96)
The cleaned-up version of
https://github.com/DutchConcepts/amigos-app/pull/92.

> Whole message view is transferred in to our own component. 
>
> All the UI is now independent from Stream's UI. Event services (like
image downloading and uploading) is decoupled from Stream. So we can
implement our own behavior (which is needed for Xcode previews) but
still benefit from Stream's services.
>
> The UI should be close to the same as `master` with some minor
improvements here and there.
Because this PR is already big I find this a valid checkpoint with every
component so reviewer can run and test it.
>
> Few things that do not work (correctly)
> - QuotedMessageView width does not scale to the message width when
there is an media item
> - Media gallery on item tap
> - Reply on deleted message view 
> - Link recognition

---------

Co-authored-by: Jarret Hoving <54537997+JarretHoving97@users.noreply.github.com>