chore: make use of skip validation flags when building iOS app (#111)
Locally for developers this shows a modal (like
[this](https://github.com/lukepistrol/SwiftLintPlugin?tab=readme-ov-file#run-on-ci)).
Obviously, our CI can't handle that, these flags will ignore it. It
would make sense that we want to be more careful which plugins we want
to add to the repository (like on web).

Futuremore, pinning the dependencies won't be a bad idea. The current
risk is minimal though.

If we really want to be strict, it seems to be possible to 'allow' each
plugin specifically but that's a bit overkill for now:
https://developer.apple.com/forums/thread/739347

https://xebia.com/blog/using-swift-macros-ci-cd-without-compromising-security/

https://forums.swift.org/t/how-do-i-get-xcode-to-programmatically-trust-a-specific-build-tool-plugin-for-ci/61331/7

At the moment, it is only needed for
[SwiftLint](https://github.com/realm/SwiftLint?tab=readme-ov-file#xcode-projects),
which lints the project when building.