feat: upgrade `@revenuecat/purchases-capacitor` (#125)
Google Play required us to use BillingClient `v6.0.1` or higher

`@revenuecat/purchases-capacitor` `v9.2.1` is the latest version to
support Capacitor `v6`. As of `v10.*` it supports Capacitor `v7`. Which
is too high for now.

Changelog for the version from `v6.1.0` up to `v9.2.1`:
https://github.com/RevenueCat/purchases-capacitor/compare/cd4220a6419d243191693a9190d22a9b7bb1c4c3...b5e20dfa972b09d5e47fd73f81cc67284ec59903#diff-06572a96a58dc510037d5efa622f9bec8519bc1beab13c9f251e97e657a9d4ed

There are no breaking changes that are relevant to us. Only important
thing was Apple's In-App Purchase Key, but we had already configured
that