fix: make channel list events work correctly (#171)
Lot's of code in composition has changed. This is primarily because we
don't need `ChannelsListViewModel` logic in the
`ChatScreen` anymore. Because when a new ViewModel creation was made,
the delegate for the ChannelsListView was pointed to a new ViewModel
instance. Which results in no UI updates happening when you navigated
back to the `ChannelListView`

This also gave me a chance to simplify the code of the composers. For
example `ExtendedStreamPlugin` does not compose the channel header view
anymore. This is done in `ChatChannelsViewControllerComposer` which
makes more sense.

also renamed a `ChatChannelsViewControllerComposer ` from `composeWith`
to `lazyCompose` so it's more clear that we open a channel and load
information on appearance when we only have a channel id to navigate to.

 which also simplify things. 
 
 
 (EDIT)

This also solves the issue of the pop-overs and or whole chat closing
when the view update by any channel event changes