fix: read extra data `activityIsActive` correctly as number and parse to int (#261)
Closes #257

Apparently this value is saved as a double. Reading and parsing to an
int directly will result in a `null` value. This line of code will then
always fallback to value `1`. Instead we should savely read and parse
the value as a Number (so that it's flexible). Then we parse it to an
int