feat(android): upgrade stream and other deps (#404)
This ships a bunch of bugfixes by Stream.

For Stream SDK to be upgraded, I also needed to upgrade Kotlin and the
AGP. Also there were some dependency conflicts that required me to also
upgrade:

- `compose-bom` from `2024.04.00` to `2025.04.00`
- `coil` from `2.7.0` to `3.2.0`

That in turn required me to make some code changes to complete the
migration. Mainly import changes and migrating from `rememberRipple` to
`ripple`